% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cells_per_dna}
\alias{cells_per_dna}
\title{Cell count per DNA}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3 rows and 3 columns.
}
\usage{
cells_per_dna
}
\description{
A data set containing the slopes to interpolate cell number from DNA
measurements.
}
\details{
\describe{
\item{cell_type}{
\code{lf}    = lung fibroblasts \cr
\code{pasmc} = pulmonary artery smooth muscle cells}
\item{volume}{volume of lysis buffer used to extract DNA}
\item{slope}{cells / ng DNA}
}
}
\keyword{datasets}
