% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copeland.2023.hypoxia.flux-package.R
\docType{package}
\name{Copeland.2023.hypoxia.flux-package}
\alias{Copeland.2023.hypoxia.flux}
\alias{Copeland.2023.hypoxia.flux-package}
\title{Copeland.2023.hypoxia.flux: Research Compendium}
\description{
Data and analysis code for "MYC overrides HIF-1α to regulate proliferating primary cell metabolism in hypoxia."
}
\author{
\strong{Maintainer}: William Oldham \email{woldham@bwh.harvard.edu} (\href{https://orcid.org/0000-0003-3029-4866}{ORCID})

}
\keyword{internal}
