% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flux_differences}
\alias{flux_differences}
\title{Flux differences}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 657 rows and 16 columns.
}
\usage{
flux_differences
}
\description{
A data table of summarized fluxes and significant differences. This also
includes data for flux differences normalized to growth rate or glucose
consumptions.
}
\details{
\describe{
\item{normalization}{
\code{none} = no normalization \cr
\code{glucose} = fluxes normalized to GLUT flux \cr
\code{growth} = fluxed normalized to biomass flux}
\item{cell_type}{
\code{lf}    = lung fibroblasts \cr
\code{pasmc} = pulmonary artery smooth muscle cells}
\item{pathway}{metabolic pathway}
\item{index}{numeric ID for a metabolic reaction}
\item{id}{name for a metabolic reaction}
\item{type}{
\code{net} = net flux \cr
\code{exch} = exchange flux}
\item{equation}{flux reaction equation}
\item{ctl}{control condition}
\item{exp}{experimental condition}
\item{flux_ctl}{flux in the control condition}
\item{lb_ctl}{lower bound of the control flux}
\item{ub_ctl}{upper bound of the control flux}
\item{flux_exp}{flux in the experimental condition}
\item{lb_exp}{lower bound of the experimental flux}
\item{ub_flux}{upper bound of the experimental flux}
\item{ratio}{ratio of the experimental flux to the control flux}
}
}
\keyword{datasets}
