% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flux_measurements}
\alias{flux_measurements}
\title{Measurements for extracellular flux calculations}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 26606 rows and 17 columns.
}
\usage{
flux_measurements
}
\description{
A data set containing the interpolated metabolite concentrations, cell counts,
and evaporation volumes for extracellular flux determinations.
}
\details{
\describe{
\item{cell_type}{
\code{lf}    = lung fibroblasts \cr
\code{pasmc} = pulmonary artery smooth muscle cells}
\item{experiment}{
\code{02}        = 0.2\% oxygen for hypoxia \cr
\code{05}        = 0.5\% oxygen for hypoxia \cr
\code{bay}       = molidustat treatment \cr
\code{05-bay}    = 0.5\% oxygen plus molidustat \cr
\code{05-simyc}  = 0.5\% oxygen plus siMYC \cr
\code{05-siphd}  = 0.5\% oxygen plus siPHD2 \cr
\code{bay-myc}   = molidustat plus MYC overexpression \cr
\code{substrate} = glucose or glutamine dropout and \if{html}{\out{<sup>}}13\if{html}{\out{</sup>}}C-glucose labeling}
\item{batch}{a group of biological replicates analyzed similarly}
\item{date}{start date of an experiment}
\item{metabolite}{name of measurement}
\item{abbreviation}{abbreviated metabolite name}
\item{detector}{
\code{fld}       = HPLC fluorescence detector \cr
\code{mwd}       = HPLC multi-wavelength detector \cr
\code{enzyme}    = enzymatic assay \cr
\code{hplc}      = OPD-derivatized HPLC detection \cr
\code{lcms}      = liquid chromatography-mass spectrometry assay \cr
\code{picogreen} = fluorescence dye labeling of DNA}
\item{type}{
\code{cells} = conditioned medium \cr
\code{empty} = unconditioned medium}
\item{oxygen}{ambient oxygen level}
\item{virus}{
\code{None} = no virus \cr
\code{YFP}  = YFP control virus \cr
\code{MYC}  = MYC overexpression virus}
\item{treatment}{
\code{None}    = no treatment \cr
\code{DMSO}    = 0.1\% DMSO \cr
\code{BAY}     = 10 μM molidustat \cr
\code{siCTL}   = non-targeting control siRNA \cr
\code{siHIF1A} = siRNA targeting HIF1A \cr
\code{siHIF2A} = siRNA targeting HIF2A \cr
\code{siMYC}   = siRNA targeting MYC \cr
\code{siPHD2}  = siRNA targeting PHD2 \cr
\code{-GLC}    = MCDB131 medium lacking glucose \cr
\code{-GLN}    = MCDB131 medium lacking glutamine \cr
\code{GLC6}    = MDCB131 medium supplemented with [U-\if{html}{\out{<sup>}}13\if{html}{\out{</sup>}}C\if{html}{\out{<sub>}}6\if{html}{\out{</sub>}}]-glucose}
\item{time}{hours}
\item{well}{denotes technical replicates in each experiment}
\item{conc}{measured in number for cells and μM for metabolites}
\item{volume}{measured in mL, extrapolated from evaporation controls}
\item{nmol}{metabolite mass}
}
}
\keyword{datasets}
