% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{k}
\alias{k}
\title{Spontaneous degradation and accumulation rates}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 54 rows and 5 columns.
}
\usage{
k
}
\description{
A data set containing the rates of significant metabolite accumulation or
degradation in cell-free medium.
}
\details{
\describe{
\item{metabolite}{name}
\item{oxygen}{ambient oxygen level}
\item{treatment}{
\code{None} = no treatment \cr
\code{DMSO} = 0.1\% DMSO \cr
\code{BAY} = 10 μM molidustat}
\item{k}{accumulation or decay rate per hour, positive values indicate decay}
}
}
\keyword{datasets}
