% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mids}
\alias{mids}
\title{Mass isotope distributions}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 44499 rows and 10 columns.
}
\usage{
mids
}
\description{
A data set containing the mass isotope distributions corrected for natural
abundance for all experiments.
}
\details{
\describe{
\item{method}{
\code{fs}  = full scan \cr
\code{sim} = selected ion monitoring}
\item{cell_type}{
\code{lf}    = lung fibroblasts \cr
\code{pasmc} = pulmonary artery smooth muscle cells}
\item{date}{start date of an experiment}
\item{tracer}{
\code{glc2} = [1,2-\if{html}{\out{<sup>}}13\if{html}{\out{</sup>}}C\if{html}{\out{<sub>}}2\if{html}{\out{</sub>}}]-glucose \cr
\code{glc6} = [U-\if{html}{\out{<sup>}}13\if{html}{\out{</sup>}}C\if{html}{\out{<sub>}}6\if{html}{\out{</sub>}}]-glucose \cr
\code{q5}   = [U-\if{html}{\out{<sup>}}13\if{html}{\out{</sup>}}C\if{html}{\out{<sub>}}5\if{html}{\out{</sub>}}]-glutamine \cr
\code{lac3} = [U-\if{html}{\out{<sup>}}13\if{html}{\out{</sup>}}C\if{html}{\out{<sub>}}3\if{html}{\out{</sub>}}]-lactate}
\item{oxygen}{ambient oxygen level}
\item{treatment}{
\code{None} = no treatment \cr
\code{DMSO} = 0.1\% DMSO \cr
\code{BAY}  = 10 μM molidustat \cr}
\item{time}{hours}
\item{metabolite}{name of measurement}
\item{isotope}{}
\item{mid}{mole fraction of the isotope}
}
}
\keyword{datasets}
