% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_fluxes}
\alias{model_fluxes}
\title{Model fluxes}
\format{
An object of class \code{grouped_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.
}
\usage{
model_fluxes
}
\description{
A data table of experimentally determined extracellular fluxes used to
constrain the metabolic model.
}
\details{
\describe{
\item{cell_type}{
\code{lf}    = lung fibroblasts \cr
\code{pasmc} = pulmonary artery smooth muscle cells}
\item{data}{
\code{metabolite} = metabolite name \cr
\code{oxygen} = ambient oxygen level \cr
\code{treatment} = \code{None}, \code{DMSO}, \code{BAY} \cr
\code{mean} = average flux from multiple biological replicates \cr
\code{se} = standard error of flux from multiple biological replicates \cr
\code{name} = corresponds to name of model reaction}
}
}
\keyword{datasets}
