% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{model_reactions}
\alias{model_reactions}
\title{Model reactions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 4 columns.
}
\usage{
model_reactions
}
\description{
A data table of model reactions formatted for INCA analysis.
}
\details{
\describe{
\item{index}{reaction identifier}
\item{name}{reaction name}
\item{equation}{equation for INCA input}
\item{pathway}{metabolic pathway}
}
}
\keyword{datasets}
