% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{qbias_correction_factors}
\alias{qbias_correction_factors}
\title{Quadrupole bias correction factors}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 124 rows and 4 columns.
}
\usage{
qbias_correction_factors
}
\description{
A data set containing experimentally determined correction factors to adjust
peak areas for isotopes detected in selected ion monitoring mode. Isotope peak
areas are multiplied by the correction factor. The corrected peak area is
subsequently used to determine the mass isotope distribution.
}
\details{
\describe{
\item{batch}{correction factors associated with specific experiments}
\item{metabolite}{name of measurement}
\item{M}{isotope}
\item{cf}{correction factor}
}
}
\keyword{datasets}
